Latest master from upstream

+++ tests/no-rpath-prebuild.sh
@@ -1,10 +1,29 @@
 #! /bin/sh -e
-SCRATCH=scratch/$(basename $0 .sh)
+set -x
+ARCH="$1"
+
+if [ -z "$ARCH" ]; then
+  ARCH=$(basename $0 .sh | sed -e 's/.*-//')
+fi
+
+SCRATCH=scratch/no-rpath-$ARCH
+
+if [ -z "$ARCH" ] || [ $ARCH = prebuild ] ; then
+  echo "Architecture required"
+  exit 1
+fi
+
+no_rpath_bin="${srcdir}/no-rpath-prebuild/no-rpath-$ARCH"
+
+if [ ! -f $no_rpath_bin ]; then
+  echo "no 'no-rpath' binary for '$ARCH' in '${srcdir}/no-rpath-prebuild'"
+  exit 1
+fi
 
 rm -rf ${SCRATCH}
 mkdir -p ${SCRATCH}
 
-cp no-rpath ${SCRATCH}/
+cp $no_rpath_bin ${SCRATCH}/no-rpath
 
 oldRPath=$(../src/patchelf --print-rpath ${SCRATCH}/no-rpath)
 if test -n "$oldRPath"; then exit 1; fi
@@ -17,5 +36,3 @@ if ! echo "$newRPath" | grep -q '/foo:/bar'; then
     echo "incomplete RPATH"
     exit 1
 fi
-
-cd ${SCRATCH} && ./no-rpath
